+++ src/test/fixtures/testSTMCubeMakefile.ts
@@ -326,7 +326,7 @@ erase: $(BUILD_DIR)/$(TARGET).elf
 # clean up
 #######################################
 clean:
-	${platform === 'win32' ? 'cmd //C rm -fR' : '-rm -fR'} $(BUILD_DIR)
+	${platform === 'win32' ? 'cmd /c rm -rf' : '-rm -fR'} $(BUILD_DIR)
 	
 #######################################
 # dependencies